Location GooglePlacesAutocomplete done
IMPORTANT: You need the API Key to have the
Finished implementing GooglePlacesAutocomplete for users when they want to change location. I removed the search and save buttons since they are no longer necessary since the user just needs to tap on the location from the dropdown. Location is saved in firestore.