Added Change Email and Change Password, Edited index.js
User can now change their email and password from within the profile settings.

To change email, they need to enter their new email they want to use, then they will receive a OTP to their old email to verify they want to change emails, then they have to open the link in their new email to change their email. All of these steps I believe should be used for the integrity of the user's account.

To Change the user's password, they need to re-enter their old password and then they can enter their new password. I will add the feature where they have a OTP to change their password later. For now, the feature works.

Edited index.js so that when the OTP is for the Email Change OTP, it would put that on the Subject. Originally, when a OTP is sent, the email would always have the subject: Password Reset OTP, but I wanted to make it so it would say Email Change OTP when the user wants to change the email. This feature is not done, but the code is there.