Added profile picture and modified Change Email
The users can now upload their profile pictures and I modified some things for the change email attribute.

When the user uploads a new profile picture, the picture will always display in their profile and profile settings in real time.

Some of the modifications for the change email feature includes not letting user's change their email to an already taken email. The only issue with the Change Email feature is that the email attribute in firestore doesn't change until the user logs in again.