Added Toggle Features for Last Name and Location
I finished making the toggle features for last name and location. I created new attributes called locationVisible and displayLocation. Use displayLocation when you want to display the location since user may not want to display their location. displayLocation will be 'N/A' when they don't want their location to be displayed.

Also changed location to displayLocation in the fetchUserData function in profile.tsx.

Also added a checkbox if user wants their location to be shared in the signup.tsx.