Added in-app notification page
In the profile page, there is a new bell button that directs the user to their notification page. When new messages are received, the bell will have a red circle to indicate users that they have a new notification.

Created a new component called notification-page.tsx.

I've create an function called addNotification inside of notificationHandler.ts. The parameters are the user's uid and the message you want to send.