Report the User Backend and Frontend
now, user matches field can be updated to "liked", "disliked", "blocked", or "reported".

if user is a reported, the matches key is changed to "reported" as well as added under "reports" field. "reports" field is a mpa to a user uid with a corresponding reason chosen and timestamp.

reports?: {
    [uid: string]: {
      reason: string;
      timestamp: any;
    };
  };

users are also able to "unblock a user" which means their interaction is restarted.